Network.TLS.Extra.Cipher

Description

cipher suite

ciphersuite_all :: [Cipher]

all encrypted ciphers supported ordered from strong to weak. this choice of ciphersuite should satisfy most normal need

ciphersuite_medium :: [Cipher]

list of medium ciphers.

ciphersuite_strong :: [Cipher]

the strongest ciphers supported.

ciphersuite_unencrypted :: [Cipher]

all unencrypted ciphers, do not use on insecure network.

ciphersuite_dhe_rsa :: [Cipher]

DHE-RSA cipher suite

ciphersuite_dhe_dss :: [Cipher]

individual ciphers

cipher_null_SHA1 :: Cipher

unencrypted cipher using RSA for key exchange and SHA1 for digest

cipher_null_MD5 :: Cipher

unencrypted cipher using RSA for key exchange and MD5 for digest

cipher_RC4_128_MD5 :: Cipher

RC4 cipher, RSA key exchange and MD5 for digest

cipher_RC4_128_SHA1 :: Cipher

RC4 cipher, RSA key exchange and SHA1 for digest

cipher_AES128_SHA1 :: Cipher

AES cipher (128 bit key), RSA key exchange and SHA1 for digest

cipher_AES256_SHA1 :: Cipher

AES cipher (256 bit key), RSA key exchange and SHA1 for digest

cipher_AES128_SHA256 :: Cipher

AES cipher (128 bit key), RSA key exchange and SHA256 for digest

cipher_AES256_SHA256 :: Cipher

AES cipher (256 bit key), RSA key exchange and SHA256 for digest

cipher_RSA_3DES_EDE_CBC_SHA1 :: Cipher

3DES cipher (168 bit key), RSA key exchange and SHA1 for digest

cipher_DHE_RSA_AES128_SHA1 :: Cipher

AES cipher (128 bit key), DHE key exchanged signed by RSA and SHA1 for digest

cipher_DHE_RSA_AES256_SHA1 :: Cipher

AES cipher (256 bit key), DHE key exchanged signed by RSA and SHA1 for digest

cipher_DHE_RSA_AES128_SHA256 :: Cipher

cipher_DHE_RSA_AES256_SHA256 :: Cipher

cipher_DHE_DSS_AES128_SHA1 :: Cipher

AES cipher (128 bit key), DHE key exchanged signed by DSA and SHA1 for digest

cipher_DHE_DSS_AES256_SHA1 :: Cipher

AES cipher (256 bit key), DHE key exchanged signed by DSA and SHA1 for digest

cipher_DHE_DSS_RC4_SHA1 :: Cipher

cipher_DHE_RSA_AES128GCM_SHA256 :: Cipher

cipher_ECDHE_RSA_AES128GCM_SHA256 :: Cipher